# bx-super-app-backend/spec/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Super APP",
    "description": "Super APP Services",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "LoginModel": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "password"]
      },
      "LogoutModel": {
        "type": "object",
        "properties": { "user_id": { "type": "string" } },
        "required": ["user_id"]
      },
      "CodeLoginModel": {
        "type": "object",
        "properties": { "code": { "type": "string" } },
        "required": ["code"]
      },
      "RefreshTokenModel": {
        "type": "object",
        "properties": { "refresh_token": { "type": "string" } },
        "required": ["refresh_token"]
      },
      "GetUserModel": {
        "type": "object",
        "properties": { "username": { "type": "string" } },
        "required": ["username"]
      },
      "ResetPasswordModel": {
        "type": "object",
        "properties": { "username": { "type": "string" } },
        "required": ["username"]
      },
      "ResetPasswordVerifyModel": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "nonce": { "type": "string" }
        },
        "required": ["username", "nonce"]
      },
      "VerifyChangePasswordModel": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "nonce": { "type": "string" },
          "new_password": { "type": "string" }
        },
        "required": ["username", "nonce", "new_password"]
      },
      "RegisterModel": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "lastname": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "avatar": { "type": "string" },
          "file": { "type": "string" }
        },
        "required": ["name", "lastname", "email", "password", "avatar", "file"]
      },
...
    }
  },
  "paths": {
    "/api/pyme2c/backend/v1/account/login": {
      "post": {
        "operationId": "AccountController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginModel" }
            }
          }
        },
        "responses": {
          "200": { "description": "User Login Success" },
          "400": { "description": "Validation Error" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/logout": {
      "post": {
        "operationId": "AccountController_logout",
        "summary": "Logout user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LogoutModel" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/code-login": {
      "post": {
        "operationId": "AccountController_codeLogin",
        "summary": "Login with code",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CodeLoginModel" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/refresh-token": {
      "post": {
        "operationId": "AccountController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenModel" }
            }
          }
        },
        "responses": {
          "200": { "description": "Refresh Token" },
          "400": { "description": "Token invalid" }
        },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/confirm-email/{username}/{code}": {
      "get": {
        "operationId": "AccountController_VerifyEmail",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "User Info" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/confirm-email/{username}/send-code": {
      "post": {
        "operationId": "AccountController_SendCode",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/userinfo": {
      "get": {
        "operationId": "AccountController_getUserInfo",
        "parameters": [],
        "responses": {
          "200": { "description": "User Info" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Account"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/pyme2c/backend/v1/account/admin/userinfo": {
      "post": {
        "operationId": "AccountController_getUserInfoByAdmin",
        "summary": "Get User Info with Admin Token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GetUserModel" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/reset-password/generate-code": {
      "post": {
        "operationId": "AccountController_resetPassword",
        "summary": "User Reset Password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordModel" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/reset-password/verify-code": {
      "post": {
        "operationId": "AccountController_resetPasswordVerify",
        "summary": "Verify code for Reset Password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordVerifyModel"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/reset-password/change-password": {
      "post": {
        "operationId": "AccountController_changeResetPassword",
        "summary": "Verify Reset Password and Change Password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyChangePasswordModel"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/register": {
      "post": {
        "operationId": "AccountController_Register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterModel" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Account"]
      }
    },
    "/api/pyme2c/backend/v1/account/bruteforce/{uid}": {
      "get": {
        "operationId": "AccountController_getUserBruteForceInfo",
        "summary": "Get Brute Force Info",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Account"]
      },
      "delete": {
        "operationId": "AccountController_resetUserBruteForceInfo",
        "summary": "Reset Brute Force/Unlock User",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Account"]
      }
    },
...
  }
}